refactor to single SponsorHandler
